requirements.txt
beautifulsoup4==4.12.3
easydict==1.13
gensim==4.3.2
grip==4.6.2
h5py==3.11.0
ipykernel==6.29.5
jieba==0.42.1
lightgbm==4.4.0
lxml==5.2.2
minkowskiengine==0.5.4
nibabel==5.2.1
ninja==1.11.1.4
nvitop==1.3.2
open3d==0.19.0
opencv-python==4.10.0.84
openpyxl==3.1.5
pip-chill==1.0.3
pysocks==1.7.1
seaborn==0.13.2
selenium==4.22.0
tensorboard==2.14.0
tensorboardx==2.6.2.2
torchaudio==0.9.0
torchvision==0.10.0+cu111
wordcloud==1.9.3
